Week_4/Day_4/requirements.txt
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
googletrans==3.1.0a0
googletranslate==2.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
psycopg2-binary==2.9.6
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.15
